Remove deprecated ai backend engine module
